+++ changelog.md
@@ -89,6 +89,7 @@ Codion Change Log
 - ComponentBuilder.listener() and consumer() added.
 - FilterTable bug fixed, sorting, column reordering and resizing key events were enabled even though the functionality had been disabled.
 - TableConditionPanel.focusEvent() added.
+- FilterColumnConditionPanel now initialized lazily.
 ### is.codion.framework.domain
 - DefaultDomain renamed DomainModel.
 - Entity.valuesEqual() instance methods renamed equalValues(), static ones removed.